+++ include/invader/error.hpp
@@ -121,5 +121,10 @@ namespace Invader {
      * This is thrown when tag data is invalid
      */
     DEFINE_EXCEPTION(InvalidTagDataException, "tag data is invalid");
+
+    /**
+     * This is thrown when a resource map was not supplied when it should have
+     */
+    DEFINE_EXCEPTION(ResourceMapRequiredException, "no resource map was supplied");
 }
 #endif